Refactored Area Loot CRUD
